requirements.txt
# Core dependencies
langgraph>=0.0.30
langchain-core>=1.3.0

# Web & Browser
playwright>=1.40.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# AI
openai>=1.0.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0